[GEOS-9936] Update GeoStyler extension
