--- src/community/geostyler/src/main/java/org/geoserver/wms/web/data/GeoStylerTabPanel.html
@@ -1,18 +1,0 @@
-<html xmlns:wicket="http://wicket.apache.org/">
-<head>
-    <title>
-        <wicket:message key="title">GeoStyler Widget for Wicket</wicket:message>
-    </title>
-</head>
-<body>
-<wicket:panel>
-    <ul class="horizontal-form">
-        <li>
-            <wicket:message key="previewOnLayer">Preview on layer: </wicket:message><a wicket:id="change.layer" href="#"></a>
-        </li>
-    </ul>
-    <div wicket:id="geoStylerDiv" />
-    <div wicket:id="dialog"/>
-</wicket:panel>
-</body>
-</html>
