+++ src/community/geostyler/src/main/java/org/geoserver/wms/web/data/js/.eslintrc
@@ -1,0 +1,43 @@
+{
+  "extends": [
+    "eslint:recommended"
+  ],
+  "plugins": [],
+  "env": {
+    "browser": true,
+    "node": true
+  },
+  "globals": {
+    "Promise": false,
+    "GeoStylerSLDParser": false,
+    "GeoStyler": false,
+    "GeoStylerWfsParser": false,
+    "React": false,
+    "ReactDOM": false,
+    "$": false
+  },
+  "rules": {
+    "indent": ["error", 2, { "SwitchCase": 1 }],
+    "linebreak-style": ["error", "unix"],
+    "quotes": ["error", "single", { "allowTemplateLiterals": true }],
+    "semi": ["error", "always"],
+    "one-var": ["error", "never"],
+    "no-confusing-arrow": "error",
+    "no-unused-vars": ["error", {
+      "ignoreRestSiblings": true
+    }],
+    "key-spacing": ["error", {
+      "beforeColon": false,
+      "afterColon": true
+    }],
+    "require-jsdoc": ["error", {
+      "require": {
+        "FunctionDeclaration": true,
+        "MethodDefinition": true,
+        "ClassDeclaration": true,
+        "ArrowFunctionExpression": true
+      }
+    }],
+    "space-infix-ops": ["error", {"int32Hint": false}]
+  }
+}
