+++ src/community/geostyler/src/main/java/org/geoserver/wms/web/data/js/geostyler-integration.css
@@ -1,0 +1,69 @@
+@font-face {
+    font-family: Aldrich;
+    src: url('https://raw.githubusercontent.com/geostyler/geostyler-demo/master/src/assets/fonts/aldrich/aldrich-v8-latin-regular.woff2') format('woff2')
+}
+
+#geoStyler {
+    border: 1px solid #a7a7a7;
+    border-radius: 3px;
+    margin-bottom: 10px;
+}
+
+#geoStylerTitle {
+    font-family: Aldrich,sans-serif;
+    color: #162162;
+    display: block;
+    font-size: 3em;
+    text-align: center;
+}
+
+#checkboxwrapper {
+    display:flex;
+    justify-content: center;
+}
+
+#checkboxwrapper label {
+    font-size: 1em;
+}
+
+#liveUpdate {
+    margin-right: 5px;
+}
+
+#geoStylerDiv {
+    text-align: center;
+}
+
+.gs-style-name-classification-row label {
+    padding-top: 7px;
+}
+
+.gs-style-name-classification-row input {
+    min-width: 250px;
+}
+
+/** handle fullscreen view */
+.fullscreen #style-context {
+    width: unset !important;
+    flex: 1;
+}
+
+.fullscreen #geoStyler {
+    flex: 1;
+}
+
+.fullscreen #styleForm #style-editor {
+    width: unset;
+    flex: 1;
+}
+
+.fullscreen .toolbar-sticky {
+    bottom: unset;
+    padding: unset;
+}
+
+.fullscreen .page-footer {
+    position: absolute;
+    bottom: 5px;
+    width: 99%;
+}
