+++ src/community/geostyler/src/main/java/org/geoserver/wms/web/data/js/package.json
@@ -7,9 +7,11 @@
     "Andreas Schmitz <schmitz@terrestris.de>",
     "André Henn <henn@terrestris.de>",
     "Kai Volland <volland@terrestris.de>",
-    "Nils Bühner <buehner@terrestris.de>"
+    "Nils Bühner <buehner@terrestris.de>",
+    "Johannes Weskamm <weskamm@terrestris.de>"
   ],
   "scripts": {
+    "lint": "eslint -c .eslintrc --ext ftl ../../../../../../../resources/org/geoserver/wms/web/data/",
     "pack:lib": "node ./npm-pack-libs.js"
   },
   "repository": {
@@ -22,7 +24,6 @@
     "Style",
     "SLD"
   ],
-  "author": "",
   "license": "GPL-3.0-or-later",
   "bugs": {
     "url": "https://github.com/geoserver/geoserver/issues"
@@ -42,5 +43,8 @@
     "geostyler-wfs-parser": "^1.0.0",
     "react": "16.x",
     "react-dom": "16.x"
+  },
+  "devDependencies": {
+    "eslint": "^7.20.0"
   }
 }
