+++ src/community/geostyler/src/main/resources/GeoServerApplication.properties
@@ -1,5 +1,5 @@
-GeoStylerTabPanel.page.title=GeoStyler
-GeoStylerTabPanel.page.description=GeoStyler style editor
+GeoStyler.page.title=GeoStyler
+GeoStyler.page.description=GeoStyler style editor
 
-GeoStylerTabPanel.title=GeoStyler
-GeoStylerTabPanel.description=GeoStyler style editor
+GeoStyler.title=GeoStyler
+GeoStyler.description=GeoStyler style editor
